refactor(cli): modularize CLI — extract parser, runner, report and main logic into dedicated files
