+++ benchmark_cherrypick/lib/cli/report/json_report.dart
@@ -1,0 +1,10 @@
+import 'report_generator.dart';
+
+class JsonReport extends ReportGenerator {
+  @override
+  List<String> get keys => [];
+  @override
+  String render(List<Map<String, dynamic>> rows) {
+    return '[\n${rows.map((r) => '  $r').join(',\n')}\n]';
+  }
+}
