+++ benchmark_cherrypick/lib/cli/report/report_generator.dart
@@ -1,0 +1,4 @@
+abstract class ReportGenerator {
+  String render(List<Map<String, dynamic>> results);
+  List<String> get keys;
+}
